fix: use default StorageClass when not specified
If the SC is not specified, or is null, K8s will pick a default SC (if
it exists) to provision dynamically the requested volumes. By now
allowing it to be null, we avoid the problem of having to detect all the
available SCs and finding the appropriate one.

This may still not provide a default on K8s environments which have more
than one default SC.

Closes #151