feat: fixes some typos on the doc
This commit adds a small changed on the doc, fixing the kind setup
commands and on the `demo-quickstart/operator-intall` added the necessary
commands to get the auto-generated user and password.

Also added a instruction about how to update the UI user and password,
and fixed patroni graphs and commands output.